+++ R/pkg.R
@@ -163,10 +163,10 @@ NULL
 
     packageStartupMessage("\ndtwclust: Setting random number generator to L'Ecuyer-CMRG (see RNGkind()).\n",
                           'To read the included vignette, type: vignette("dtwclust").\n',
-                          'Please see news(package = "dtwclust") for important information!\n')
+                          'Please see news(package = "dtwclust") for important information.\n')
 
     if (grepl("\\.9000$", utils::packageVersion("dtwclust")))
-        packageStartupMessage("This is a developer version of 'dtwclust'.")
+        packageStartupMessage("This is a developer version of 'dtwclust'. Using devtools::test() is currently broken.")
 }
 
 .onUnload <- function(libpath) {