+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Version 3.1.1
 * Seeds are now set when calling TADPole through `dtwclust` with a parallel backend and multiple values of `k`, in case the centroid function has randomness associated.
-* Added a new function `tsclust` that should be functionally equivalent to `dtwclust` but is hopefully more coherent in general.
+* Added a new experimental function `tsclust` that should be functionally equivalent to `dtwclust` but is hopefully more coherent in general.
 
 ## Version 3.1.0
 * Fixed subsetting of multivariate plots